style: add file for each asset select item component =)
