+++ src/module/wallet/component/input/WalletAssetSelect/InnerAssetSelect/TokenSelectItemList/TokenSelectItem.tsx
@@ -1,8 +1,7 @@
-import useGetTokens from "module/token/query/useGetTokens";
 import { AssetType } from "module/wallet/wallet.types";
 import { Token } from "near-peersyst-sdk";
-import BaseTokenSelectItem from "./BaseTokenSelectItem";
-import { useAssetSelect } from "../hook/useAssetSelect";
+import { useAssetSelect } from "../../hook/useAssetSelect";
+import BaseTokenSelectItem from "../BaseTokenSelectItem";
 
 export interface TokenSelectItemProps {
     token: Token;
@@ -21,11 +20,3 @@ export const TokenSelectItem = ({ token }: TokenSelectItemProps) => {
 
     return <BaseTokenSelectItem onPress={handleOnPress} units={symbol} balance={balance} token={token} />;
 };
-
-const TokenSelectItemlist = (): JSX.Element => {
-    const { index } = useAssetSelect();
-    const { data: tokens = [] } = useGetTokens(index);
-    return <>{tokens.length > 0 && tokens.map((token, index) => <TokenSelectItem key={index} token={token} />)}</>;
-};
-
-export default TokenSelectItemlist;
